kernel: eliminate casts where possible
This can allow the compiler to detect more errors.

In most cases, using the "z" length modifier in format strings (for
size_t and sssize_t) did the trick.  "t" was used for ptrdiff_t.

In ipsec_rcv_ah_checks(): changed grouping to avoid overflow.

In pfkey_msg_build(): simplify a contorted memcpy.

+++ linux/net/ipsec/ipsec_ah.c
@@ -95,9 +95,9 @@ enum ipsec_rcv_value ipsec_rcv_ah_checks(struct ipsec_rcv_state *irs,
 	    ((AHHMAC_HASHLEN + AHHMAC_RPLLEN) >> 2)) {
 		KLIPS_PRINT(debug_rcv & DB_RX_INAU,
 			    "klips_debug:ipsec_rcv: "
-			    "bad authenticator length %ld, expected %lu from %s.\n",
-			    (long)(irs->protostuff.ahstuff.ahp->ah_hl << 2),
-			    (unsigned long) sizeof(struct ahhdr),
+			    "bad authenticator length %ld, expected %zu from %s.\n",
+			    (long)irs->protostuff.ahstuff.ahp->ah_hl << 2,
+			    sizeof(struct ahhdr),
 			    irs->ipsaddr_txt);
 		if (irs->stats)
 			irs->stats->rx_errors++;
